backend/requirements.txt
Flask
flask-cors
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
gunicorn